# remove dupe validation in WriteRelationships