I noticed in flamegraphs that validation was taking a chunk of the request time on large WriteRelationshipsRequest payloads, and that it was happening two times:
- one through the validation middleware
- one in `permissionServer.WriteRelationships` to detect dupes

<img width="1486" alt="Screenshot 2022-09-05 at 17 56 10" src="https://user-images.githubusercontent.com/6774726/188628170-d1011580-113b-43fc-b09a-c23a54476cb7.png">

This was a regression from https://github.com/authzed/spicedb/pull/796

The commit improves the performance in various ways:
- removes the duplicated validation
- reduces CPU time to generate a v1.Relationship string in 85%
- reduces allocations to generate a v1.Relationship string in 100% (no allocations)

Output from `benchstat` from the new `StringRelationship` function. FWIW this is not going to have a dramatic impact on the request path, but given I was removing the duped validation, I decided to have a closer look for any low hanging fruit.

```
name                   old time/op    new time/op    delta
StringRelationship-16     516ns ± 2%      77ns ± 2%   -85.08%  (p=0.008 n=5+5)

name                   old alloc/op   new alloc/op   delta
StringRelationship-16      192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)

name                   old allocs/op  new allocs/op  delta
StringRelationship-16      11.0 ± 0%       0.0       -100.00%  (p=0.008 n=5+5)
```

## Bonus: disable server logging during test

I also disabled server logs during tests in  https://github.com/authzed/spicedb/commit/62f1401a0075db3beceffcf3f4337916a986b643
Makes tests slightly faster with `-json` and with a less cluttered output